modules: make .module_license read-only
Currently .module_license is set writable (that is, the section has the
SHF_WRITE flag set) in the module's ELF headers.  This probably never
actually matters, but it can't possibly be correct.

This patch sets that data as "const", which causes that flag not to be
set.

Signed-off-by: Peter Jones <pjones@redhat.com>
(cherry picked from commit f6563e15bb490bb76a1a95cd3648fe03d1134d14)
Signed-off-by: Jan Setje-Eilers <jan.setjeeilers@oracle.com>