modules: strip .llvm_addrsig sections and similar.
Currently grub modules built with clang or gcc have several sections
which we don't actually need or support.

We already have a list of section to skip in genmod.sh, and this patch
adds the following sections to that list (as well as a few newlines):

.note.gnu.property
.llvm*

Note that the glob there won't work without a new enough linker, but the
failure is just reversion to the status quo, so that's not a big problem.

Signed-off-by: Peter Jones <pjones@redhat.com>
(cherry picked from commit 0f66524e94d3c4f4d669d75c2122b0f1036776ea)
Signed-off-by: Jan Setje-Eilers <jan.setjeeilers@oracle.com>